Moves objects to package
