+++ m3/Packages/com.oldbard.services.match3/Runtime/Grid/Data/GameConfig.cs
@@ -1,4 +1,4 @@
-﻿using OldBard.Services.Match3.Grid.Views;
+﻿using OldBard.Services.Match3.Grid.View;
 using UnityEngine;
 
 namespace OldBard.Services.Match3.Grid.Data
