+++ m3/Packages/com.oldbard.services.match3/Runtime/Grid/Factories/TileInstanceFactory.cs
@@ -1,6 +1,6 @@
 ﻿using OldBard.Services.Match3.Grid;
 using OldBard.Services.Match3.Grid.Data;
-using OldBard.Services.Match3.Grid.Views;
+using OldBard.Services.Match3.Grid.View;
 using UnityEngine;
 using UnityEngine.Pool;
