+++ m3/Packages/com.oldbard.services.match3/Runtime/Grid/com.oldbard.services.match3.grid.asmdef
@@ -2,7 +2,7 @@
   "name": "com.oldbard.services.match3.grid",
   "rootNamespace": "OldBard.Services.Match3",
   "references": [
-	"Unity"
+    "Unity.TextMeshPro"
   ],
   "includePlatforms": [],
   "excludePlatforms": [],
